+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+* Upgrade the `org.realityforge.proton` artifact to version `0.68`.
+* Migrate `javapoet` to `com.palantir.javapoet:javapoet:jar:0.14.0` which is still maintained.
+
 ### [v0.239](https://github.com/arez/arez/tree/v0.239) (2026-04-28) · [Full Changelog](https://github.com/arez/arez/compare/v0.238...v0.239)
 
 Changes in this release: